SceneManagerの内容をGameに総入れ替え
