+++ main.cpp
@@ -102,37 +102,15 @@ int WINAPI WinMain(HINSTANCE, HINSTANCE, LPSTR, int)
 
 	// インスタンス化
 	Game* game = new Game();
-    SceneManager* sceneManager = new SceneManager();
-	
-	// 全オブジェクトのnew生成
-	//game->Create();
-
-	// ゲームの初期化
-	//game->Initialize();
-	//game->InitializeGameStart();
-    //sceneManager->Initialize();
 
 	// エスケープキーが押されるかウインドウが閉じられるまでループ
 	while (ProcessMessage() == 0 && CheckHitKey(KEY_INPUT_ESCAPE) == 0)
 	{
 		// FPS処理
 		auto _prevTime = GetNowHiPerformanceCount();
-		
-		// ゲームの更新
-		//game->UpdateGame();
-        //sceneManager->Update();   // シーン管理
 
-		// 画面を初期化する
-		//ClearDrawScreen();
-
-        sceneManager->GameLoop();
-        
-		// ゲームの描画
-		//game->DrawGame();
-        //sceneManager->Draw(); // シーン管理
-
-		// 裏画面の内容を表画面に反映させる
-		//ScreenFlip();
+        // ゲームループ
+        game->GameLoop();
 
 		// FPS処理
    		auto _afterTime = GetNowHiPerformanceCount();
@@ -154,9 +132,6 @@ int WINAPI WinMain(HINSTANCE, HINSTANCE, LPSTR, int)
 			break;
 		}
 	}
-	
-	// 生成したものを削除
-	//game->DeleteGame();
 
 	// Effekseerを終了する。
 	//Effkseer_End();
